requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pytest-django==4.7.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2==2.9.9
python-dotenv==1.0.0
django-redis==5.4.0
xmltodict==0.13.0
redis==5.0.1
requests==2.31.0